#173 reorganize filehub tests, add filehub money related filehub methods
