+++ src/listen/test_filehub/test_filehub__core.py
@@ -6,21 +6,16 @@ from src._road.road import (
 )
 from src._road.finance import default_planck_if_none
 from src._road.jaar_config import (
+    get_atoms_folder,
     get_test_reals_dir,
     get_test_real_id,
     get_rootpart_of_econ_dir,
 )
 from src.agenda.agenda import agendaunit_shop
-from src.listen.filehub import (
-    get_econ_path,
-    FileHub,
-    filehub_shop,
-    get_nox_type_set,
-)
+from src.listen.filehub import FileHub, filehub_shop, get_econ_path, get_nox_type_set
 from src.listen.examples.examples import get_agenda_with_4_levels
 from src.listen.examples.listen_env import (
-    get_texas_filehub,
-    get_listen_temp_env_dir,
+    get_listen_temp_env_dir as env_dir,
     env_dir_setup_cleanup,
 )
 from pytest import raises as pytest_raises
@@ -85,6 +80,42 @@ def test_FileHub_Exists():
     assert x_filehub.planck is None
 
 
+def test_filehub_shop_ReturnsCorrectObj():
+    # GIVEN
+    x_reals_dir = "src/real/examples"
+    x_real_id = "music"
+    sue_text = "Sue"
+    x_road_delimiter = "/"
+    x_planck = 3
+
+    # WHEN
+    x_filehub = filehub_shop(
+        x_reals_dir, x_real_id, sue_text, None, None, x_road_delimiter, x_planck
+    )
+
+    # THEN
+    assert x_filehub.reals_dir == x_reals_dir
+    assert x_filehub.real_id == x_real_id
+    assert x_filehub.person_id == sue_text
+    assert x_filehub.road_delimiter == x_road_delimiter
+    assert x_filehub.planck == x_planck
+
+    assert x_filehub.real_dir() == f"{x_reals_dir}/{x_real_id}"
+    assert x_filehub.persons_dir() == f"{x_filehub.real_dir()}/persons"
+    assert x_filehub.person_dir() == f"{x_filehub.persons_dir()}/{sue_text}"
+    assert x_filehub.econs_dir() == f"{x_filehub.person_dir()}/econs"
+    assert x_filehub.quarks_dir() == f"{x_filehub.person_dir()}/quarks"
+    assert x_filehub.duty_dir() == f"{x_filehub.person_dir()}/duty"
+    assert x_filehub.work_dir() == f"{x_filehub.person_dir()}/work"
+    assert x_filehub.atoms_dir() == f"{x_filehub.person_dir()}/{get_atoms_folder()}"
+    assert x_filehub.duty_file_name() == f"{sue_text}.json"
+    x_duty_file_path = f"{x_filehub.duty_dir()}/{x_filehub.duty_file_name()}"
+    assert x_filehub.duty_file_path() == x_duty_file_path
+    assert x_filehub.work_file_name() == f"{sue_text}.json"
+    x_workpath = f"{x_filehub.work_dir()}/{x_filehub.work_file_name()}"
+    assert x_filehub.work_path() == x_workpath
+
+
 def test_filehub_shop_ReturnsCorrectObjWhenEmpty():
     # GIVEN
     sue_text = "Sue"
@@ -122,199 +153,100 @@ def test_filehub_shop_ReturnsCorrectObjWhenEmpty():
     assert sue_filehub.treasury_db_path() == treasury_file_path
 
 
-def test_FileHub_create_econ_dir_if_missing_CreatesDirectory(env_dir_setup_cleanup):
+def test_filehub_shop_RaisesErrorIf_person_id_Contains_road_delimiter():
     # GIVEN
-    sue_text = "Sue"
-    nation_text = "nation-state"
-    nation_road = create_road(root_label(), nation_text)
-    usa_text = "USA"
-    usa_road = create_road(nation_road, usa_text)
-    texas_text = "Texas"
-    texas_road = create_road(usa_road, texas_text)
-    temp_env_dir = get_listen_temp_env_dir()
-    sue_filehub = filehub_shop(temp_env_dir, None, sue_text, texas_road)
-    assert os_path_exists(sue_filehub.econ_dir()) is False
-
-    # WHEN
-    sue_filehub.create_econ_dir_if_missing()
-
-    # THEN
-    assert os_path_exists(sue_filehub.econ_dir())
-
-
-def test_FileHub_save_role_agenda_CorrectlySavesFile(env_dir_setup_cleanup):
-    # GIVEN
-    sue_text = "Sue"
-    nation_text = "nation-state"
-    nation_road = create_road(root_label(), nation_text)
-    usa_text = "USA"
-    usa_road = create_road(nation_road, usa_text)
-    texas_text = "Texas"
-    texas_road = create_road(usa_road, texas_text)
-    temp_env_dir = get_listen_temp_env_dir()
-    sue_filehub = filehub_shop(temp_env_dir, None, sue_text, texas_road)
-    bob_text = "Bob"
-    bob_agenda = get_agenda_with_4_levels()
-    bob_agenda.set_owner_id(bob_text)
-    assert os_path_exists(sue_filehub.role_path(bob_text)) is False
-
-    # WHEN
-    sue_filehub.save_role_agenda(bob_agenda)
-
-    # THEN
-    assert os_path_exists(sue_filehub.role_path(bob_text))
-
-
-def test_FileHub_role_file_exists_ReturnsCorrectBool(env_dir_setup_cleanup):
-    # GIVEN
-    sue_text = "Sue"
-    nation_text = "nation-state"
-    nation_road = create_road(root_label(), nation_text)
-    usa_text = "USA"
-    usa_road = create_road(nation_road, usa_text)
-    texas_text = "Texas"
-    texas_road = create_road(usa_road, texas_text)
-    temp_env_dir = get_listen_temp_env_dir()
-    sue_filehub = filehub_shop(temp_env_dir, None, sue_text, texas_road)
-    bob_text = "Bob"
-    bob_agenda = get_agenda_with_4_levels()
-    bob_agenda.set_owner_id(bob_text)
-    assert sue_filehub.role_file_exists(bob_text) is False
-
-    # WHEN
-    sue_filehub.save_role_agenda(bob_agenda)
-
-    # THEN
-    assert sue_filehub.role_file_exists(bob_text)
-
-
-def test_FileHub_get_role_agenda_OpensFile(env_dir_setup_cleanup):
-    # GIVEN
-    sue_text = "Sue"
-    nation_text = "nation-state"
-    nation_road = create_road(root_label(), nation_text)
-    usa_text = "USA"
-    usa_road = create_road(nation_road, usa_text)
-    texas_text = "Texas"
-    texas_road = create_road(usa_road, texas_text)
-    temp_env_dir = get_listen_temp_env_dir()
-    sue_filehub = filehub_shop(temp_env_dir, None, sue_text, texas_road)
-    bob_text = "Bob"
-    bob_agenda = get_agenda_with_4_levels()
-    bob_agenda.set_owner_id(bob_text)
-    sue_filehub.save_role_agenda(bob_agenda)
+    slash_text = "/"
+    bob_text = f"Bob{slash_text}Sue"
 
     # WHEN / THEN
-    assert sue_filehub.get_role_agenda(bob_text).get_dict() == bob_agenda.get_dict()
+    with pytest_raises(Exception) as excinfo:
+        filehub_shop(None, None, person_id=bob_text, road_delimiter=slash_text)
+    assert (
+        str(excinfo.value)
+        == f"'{bob_text}' needs to be a RoadNode. Cannot contain delimiter: '{slash_text}'"
+    )
 
 
-def test_FileHub_delete_role_file_DeletesAgendaFile(env_dir_setup_cleanup):
-    # GIVEN
-    texas_filehub = get_texas_filehub()
-    sue_agenda = get_agenda_with_4_levels()
-    sue_text = sue_agenda._owner_id
-    texas_filehub.save_role_agenda(sue_agenda)
-    print(f"{texas_filehub.role_path(sue_text)=}")
-    role_path = texas_filehub.role_path(sue_text)
-    assert os_path_exists(role_path)
-
-    # WHEN
-    texas_filehub.delete_role_file(sue_text)
-
-    # THEN
-    assert os_path_exists(role_path) is False
-
-
-def test_FileHub_save_job_agenda_CorrectlySavesFile(env_dir_setup_cleanup):
+def test_FileHub_save_file_duty_CorrectlySavesFile(env_dir_setup_cleanup):
     # GIVEN
     sue_text = "Sue"
-    nation_text = "nation-state"
-    nation_road = create_road(root_label(), nation_text)
-    usa_text = "USA"
-    usa_road = create_road(nation_road, usa_text)
-    texas_text = "Texas"
-    texas_road = create_road(usa_road, texas_text)
-    temp_env_dir = get_listen_temp_env_dir()
-    sue_filehub = filehub_shop(temp_env_dir, None, sue_text, texas_road)
-    bob_text = "Bob"
-    bob_agenda = get_agenda_with_4_levels()
-    bob_agenda.set_owner_id(bob_text)
-    assert os_path_exists(sue_filehub.job_path(bob_text)) is False
+    sue_filehub = filehub_shop(env_dir(), None, sue_text)
+    assert os_path_exists(sue_filehub.duty_file_path()) is False
 
     # WHEN
-    sue_filehub.save_job_agenda(bob_agenda)
+    sue_filehub.save_file_duty(file_text="fooboo", replace=True)
 
     # THEN
-    assert os_path_exists(sue_filehub.job_path(bob_text))
+    assert os_path_exists(sue_filehub.duty_file_path())
 
 
-def test_FileHub_job_file_exists_ReturnsCorrectBool(env_dir_setup_cleanup):
+def test_FileHub_duty_file_exists_ReturnsCorrectBool(env_dir_setup_cleanup):
     # GIVEN
     sue_text = "Sue"
-    nation_text = "nation-state"
-    nation_road = create_road(root_label(), nation_text)
-    usa_text = "USA"
-    usa_road = create_road(nation_road, usa_text)
-    texas_text = "Texas"
-    texas_road = create_road(usa_road, texas_text)
-    temp_env_dir = get_listen_temp_env_dir()
-    sue_filehub = filehub_shop(temp_env_dir, None, sue_text, texas_road)
-    bob_text = "Bob"
-    bob_agenda = get_agenda_with_4_levels()
-    bob_agenda.set_owner_id(bob_text)
-    assert sue_filehub.job_file_exists(bob_text) is False
+    sue_filehub = filehub_shop(env_dir(), None, sue_text)
+    assert sue_filehub.duty_file_exists() is False
 
     # WHEN
-    sue_filehub.save_job_agenda(bob_agenda)
+    sue_filehub.save_file_duty(file_text="fooboo", replace=True)
 
     # THEN
-    assert sue_filehub.job_file_exists(bob_text)
+    assert sue_filehub.duty_file_exists()
 
 
-def test_FileHub_get_job_agenda_OpensFile(env_dir_setup_cleanup):
+def test_FileHub_open_file_duty_OpensFile(env_dir_setup_cleanup):
     # GIVEN
     sue_text = "Sue"
-    nation_text = "nation-state"
-    nation_road = create_road(root_label(), nation_text)
-    usa_text = "USA"
-    usa_road = create_road(nation_road, usa_text)
-    texas_text = "Texas"
-    texas_road = create_road(usa_road, texas_text)
-    temp_env_dir = get_listen_temp_env_dir()
-    sue_filehub = filehub_shop(temp_env_dir, None, sue_text, texas_road)
-    bob_text = "Bob"
-    bob_agenda = get_agenda_with_4_levels()
-    bob_agenda.set_owner_id(bob_text)
-    sue_filehub.save_job_agenda(bob_agenda)
+    sue_filehub = filehub_shop(env_dir(), None, sue_text)
+    example_text = "fooboo"
+    sue_filehub.save_file_duty(example_text, replace=True)
 
     # WHEN / THEN
-    assert sue_filehub.get_job_agenda(bob_text).get_dict() == bob_agenda.get_dict()
+    assert sue_filehub.open_file_duty() == example_text
 
 
-def test_FileHub_delete_job_file_DeletesAgendaFile(env_dir_setup_cleanup):
+def test_FileHub_save_file_work_CorrectlySavesFile(env_dir_setup_cleanup):
     # GIVEN
-    texas_filehub = get_texas_filehub()
-    sue_agenda = get_agenda_with_4_levels()
-    sue_text = sue_agenda._owner_id
-    texas_filehub.save_job_agenda(sue_agenda)
-    print(f"{texas_filehub.job_path(sue_text)=}")
-    job_path = texas_filehub.job_path(sue_text)
-    assert os_path_exists(job_path)
+    sue_text = "Sue"
+    sue_filehub = filehub_shop(env_dir(), None, sue_text)
+    assert os_path_exists(sue_filehub.work_path()) is False
 
     # WHEN
-    texas_filehub.delete_job_file(sue_text)
+    sue_filehub.save_file_work(file_text="fooboo", replace=True)
 
     # THEN
-    assert os_path_exists(job_path) is False
+    assert os_path_exists(sue_filehub.work_path())
+
+
+def test_FileHub_work_file_exists_ReturnsCorrectBool(env_dir_setup_cleanup):
+    # GIVEN
+    sue_text = "Sue"
+    sue_filehub = filehub_shop(env_dir(), None, sue_text)
+    assert sue_filehub.work_file_exists() is False
+
+    # WHEN
+    sue_filehub.save_file_work(file_text="fooboo", replace=True)
+
+    # THEN
+    assert sue_filehub.work_file_exists()
+
+
+def test_FileHub_open_file_work_OpensFile(env_dir_setup_cleanup):
+    # GIVEN
+    sue_text = "Sue"
+    sue_filehub = filehub_shop(env_dir(), None, sue_text)
+    example_text = "fooboo"
+    sue_filehub.save_file_work(example_text, replace=True)
+
+    # WHEN / THEN
+    assert sue_filehub.open_file_work() == example_text
 
 
 def test_FileHub_save_duty_agenda_CorrectlySavesFile(env_dir_setup_cleanup):
     # GIVEN
     sue_agendaunit = get_agenda_with_4_levels()
     sue_text = sue_agendaunit._owner_id
-    env_dir = get_listen_temp_env_dir()
     real_id = root_label()
-    sue_filehub = filehub_shop(env_dir, real_id, sue_text, None)
+    sue_filehub = filehub_shop(env_dir(), real_id, sue_text, None)
 
     print(f"{sue_filehub.duty_file_path()=}")
     assert sue_filehub.duty_file_exists() is False
@@ -331,9 +263,9 @@ def test_FileHub_save_duty_agenda_RaisesErrorWhenAgenda_work_id_IsWrong(
 ):
     # GIVEN
     sue_text = "Sue"
-    env_dir = get_listen_temp_env_dir()
+
     real_id = root_label()
-    sue_filehub = filehub_shop(env_dir, real_id, sue_text, None)
+    sue_filehub = filehub_shop(env_dir(), real_id, sue_text, None)
 
     # WHEN / THEN
     yao_text = "yao"
@@ -355,8 +287,7 @@ def test_FileHub_get_duty_agenda_OpensFile(env_dir_setup_cleanup):
     usa_road = create_road(nation_road, usa_text)
     texas_text = "Texas"
     texas_road = create_road(usa_road, texas_text)
-    temp_env_dir = get_listen_temp_env_dir()
-    sue_filehub = filehub_shop(temp_env_dir, None, sue_text, texas_road)
+    sue_filehub = filehub_shop(env_dir(), None, sue_text, texas_road)
     sue_filehub.save_duty_agenda(sue_agendaunit)
 
     # WHEN / THEN
@@ -367,9 +298,9 @@ def test_FileHub_save_work_agenda_CorrectlySavesFile(env_dir_setup_cleanup):
     # GIVEN
     sue_agendaunit = get_agenda_with_4_levels()
     sue_text = sue_agendaunit._owner_id
-    env_dir = get_listen_temp_env_dir()
+
     real_id = root_label()
-    sue_filehub = filehub_shop(env_dir, real_id, sue_text, None)
+    sue_filehub = filehub_shop(env_dir(), real_id, sue_text, None)
 
     print(f"{sue_filehub.work_path()=}")
     assert sue_filehub.work_file_exists() is False
@@ -391,8 +322,7 @@ def test_FileHub_get_work_agenda_OpensFile(env_dir_setup_cleanup):
     usa_road = create_road(nation_road, usa_text)
     texas_text = "Texas"
     texas_road = create_road(usa_road, texas_text)
-    temp_env_dir = get_listen_temp_env_dir()
-    sue_filehub = filehub_shop(temp_env_dir, None, sue_text, texas_road)
+    sue_filehub = filehub_shop(env_dir(), None, sue_text, texas_road)
     sue_filehub.save_work_agenda(sue_agendaunit)
 
     # WHEN / THEN
@@ -404,9 +334,9 @@ def test_FileHub_save_work_agenda_RaisesErrorWhenAgenda_work_id_IsWrong(
 ):
     # GIVEN
     sue_text = "Sue"
-    env_dir = get_listen_temp_env_dir()
+
     real_id = root_label()
-    sue_filehub = filehub_shop(env_dir, real_id, sue_text, None)
+    sue_filehub = filehub_shop(env_dir(), real_id, sue_text, None)
 
     # WHEN / THEN
     yao_text = "yao"
